#3644 fixed jdbc connection parameters such that daylight savings is no longer a problem.
Needs testing on deployment.

Background:
There was a bug with the previous version of the jconnector that mishandled the tz specified in the string (and just used the local tz regardless). At that time, it seemed that using gmtMillis for timestamps/datetimes alleviated the problem. However, with the current jconnector 5.1.11 (and the deployed version 5.1.10) gmtMillis can be incorrect (e.g., off by an hour) when fetching data. So by  removing (or setting to false) this property now does the right thing.