#7329 Fix to the NACK problem - GSX Confirmation ID is now not considered anywhere in code for a NACK
Also possible fix for the status check - it appears a non-used Job field: the gsxRepairId was used to perform the checks - this is a recent change.
Since the field has not been populated for a long time (if ever), the status check was never even attempted.
Removed the field altogether from RemedyConnect model, UI and code base.
Also fixed the menu tab highlight problem gratis where logic is inverted.