apply r11401 to trunk:
fixed a problem when a database field was fetched to iterate through a list of records. The field is related to the list, but for safety it is better to use the list count itself.

also added update code to the triggers on server, this will make sure that the session count stored in course class matches the actual number of sessions.

Olga please review this commit, then if it is ok port to trunk, thanks. #7743