#15951 A complete refactoring of the AVTEMISS 7 export. It passes unit tests, but is likely to have a number of problems in this state.
I am authorising the addition of this new feature late in the beta cycle since this code is completely isolated and does not affect any existing functionality. If this is not ready in time for 4.1, we'll just disable the menu items.

Natalie, this work has raised a number of questions I need to discuss with you.