#0000 Add gradle wrapper. You don't need to know anything about this, except instead of running:
    gradle test

Now you should run:

    ./gradlew test

The long term advantage to this is that the gradle version is now controlled in svn, so all developers will use the exact same version always. The short term reason is that I want to get the Sonar plugin working again and it is much better in gradle 1.5.