- #17691
Refactor the app structure to not build two separate builds (app and website), since we will only be creating the actual app.

Completely remove the Jade dependency and use dynamically loaded HTML templates.

Create a separate API layer that will communicate with the API and serve the data to the angular Data service.

Create the Login partials and controller, and integrate it with the API and Data service.

Improve Sass structure, by separating into multiple components. This will also avoid conflicts.

Create the Talents partial and controller, and integrate it with the API and Data service.

Create the user.json and talents.json mocks, to mimic API responses.

Integrate Karma test runner, with Mocha-based tests and create a very simple test for the Login, to make sure the integration works.