#19341 I returned back "jnlp. prefix" implementation because all other approaches (to use template and denote dynamic parameters by *, to pass these parameters as main class arguments) have not gave positive results.
prefix jnlp. was added for all system properties in jnlp files.
 SystemUtils was introduced to allow to get system properties with and without jnlp prefix.
The changes was introduced because Starting from jvm 7 Upgrade 45 we should use the prefix for properties that we use to configure the application from the JNLP file. 
The workaround was took from:<br/>http://stackoverflow.com/questions/19400725/with-java-update-7-45-the-system-properties-no-more-set-from-jnlp-tag-property/19405626#19405626